feat: reorganize constants and services; move DOMUtils and constants to new files for better structure
