+++ src/content/services/index.ts
@@ -1,3 +1,4 @@
 export { observeDOM } from './DOMObserver'
 export { isRedminePage, findTextareas, hideToolbars, isTextareaInContext } from './RedmineService'
 export { canProcess, processTextarea, cleanupTextarea, cleanupAll } from './TextareaProcessor'
+export { DOMUtils } from './domUtils'
